Added in ThemeController: * getAll() - find all information in theme * an Optional<> for finding themes by id
Altered in ArtistController:
* Removed unused import statements
* Changed the path variable from int to long

+++ backend/src/main/java/no/kristiania/munchMoment/endpoints/ArtistController.java
@@ -2,12 +2,9 @@ package no.kristiania.munchMoment.endpoints;
 
 import no.kristiania.munchMoment.database.ArtistRepository;
 import no.kristiania.munchMoment.entities.Artist;
-import no.kristiania.munchMoment.entities.Painting;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.HttpMediaTypeNotAcceptableException;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Optional;
@@ -21,7 +18,7 @@ public class ArtistController {
     ArtistRepository artistRepository;
 
     @GetMapping("/{id}")
-    public ResponseEntity<Artist> getArtist(@PathVariable int id) {
+    public ResponseEntity<Artist> getArtist(@PathVariable long id) {
         Optional<Artist> optionalArtist = artistRepository.findById(id);
         return optionalArtist
                 .map(artist -> ResponseEntity.status(HttpStatus.OK).body(artist))
